Added support for XML (I think)
I have never worked with a VSCode extension before, but working with what I already know and some of the existing code here I added a snippet that should detect XML comments.
https://github.com/rioj7/remove-comments/issues/28
There might be more that needs to be done, I do not know how I can test this my build to see if it does work or not, but this is a start atleast.